Fix scrollbar issue #1
Issue #1 stated that the scrollbar does not show up in blog pages or
project pages. However, after investigation and using the Chrome dev
tools, it was found that the parent styles inherited by the index.js is
loaded even for other pages. Since the index.module.css sets the
overflow-y of html, body, and #___gatsby to "hidden", it will hide the
scrollbar fro all pages. However, it should be set to "auto" instead to
allow the scrollbar to be dynamically displayed.