Fix post out of order issue
GraphQL returns next as the node after the current and previous as the
node before However, our logical layout is next <-> current <-> previous
for posts so we have to sort the posts by ASC dates so that it will follow
this logical layout. This change has to happen when creating the pages.