Add preview loading image carousel component
Using static queries, we can build a carousel component that will
retrieve all the images for posts and subsequently filter the images by
relativeDirectory to produce only the images related to the current
project selected.

Due to limitations of static query, we are unable to use variables in
the GraphQL query, which means that we have to query for ALL images
first before we can filter down what we actually require. If the
Gatsby.js team decides to add variable support for static query, the
performance of this component would increase significantly.